requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
tensorflow>=2.8.0
keras>=2.8.0
joblib>=1.1.0
xgboost>=1.5.0
streamlit>=1.28.0
plotly>=5.15.0
kafka-python>=2.0.2 